move: Double formatting 전역 extension으로 이동
